Make this async (at the top level)
We will do network stuff, specifically HTTP. This lends itself to doing
things async.

We decided to use tokio, because that's more or less our default runtime
for these kinds of systems. For now, we only need the basic runtime. We
decide to go for the single-threaded one because the service will
(hopefully) be quite lightweight.