deps: add warp and serde
We intend to write a small HTTP REST service. For the HTTP part, we
decided to use `warp` since it appears to be the simplest reasonable
option for what we want to do.

Since a REST API will involve some JSON, we also pull in `serde`. We
don't actually need to pull that in right now, as `warp` already offers
some convenience functionality w.r.t. JSON (that's why don't  pull in
`serde_json`). However, the interfaces still expose some `serde` traits,
and we'll likely have to name them at some point.