Add a convenience trait for transforming stuff into a Reply
We intend to write a small HTTP service using `warp`. It's purpose is to
expose some information on local GPUs to other services via a REST API.
This will, in almost all cases, mean that we get some value which needs
to be serialized. Additionally, when querying those values, we may
encounter various errors that we need to handle. In short, we need to
generate an appropriate `warp::Reply` from a `Result` in quite a few
places.

While this may be somewhat straght-forward, it involves some amount of
boiler-plate code. In order to minimize that, this change introduces a
convenience trait with an impl for `Result`.