Add a fn for performing some operation with a device
We plan to offer some operations on devices under a common subpath
`/v1/device/<id>/`. Ideally, we would construct a `Filter` which yields
the `nvml::Device` for the `id` or, if the device does not exist, a
rejection which ultimately results in a `404 Not Found`.

However, `nvml::Device` is lifetime-dependent on `nvml::Nvml`. Thus, we
cannot construt such a `Filter`. Instead, we need to fetch the device in
the operation. The fact that we may need to reject the request based on
the id, however, imposes some boiler-plate. We do not want to take a
short-cut here and return a `Reply` directly, because this may obstruct
recovering and thus some composition of `Filter`s.

So we decided to outsource as much of that boilder-plate as we can, by
supplying a function which takes an operation and gets the `Device` for
it. If we were to return an `Fn` we could reduce the boiler-plate even
further, but that would require an unconfortable amount of explicit
typing, in part because `warp::FilterBase` is private.