impl Replyify for NvmlError
`Replyify` was meant to make handling of `Result`s more convenient
inside endpoints/operations by offering a shortcut from a `Result`
directly to JSON output. However, we will likely want to bail out
using the Try-operator `?` but continue processing the result. In
such situations, we just want some form of `Result::map_err` for
transforming just the error into a `warp::Reply`.

Unfortunately, `Replyify` doesn't lend itself to an additional fn
only covering the `Err` variant of a `Result` because it's completely
agnostic. However, this means we can easily impl it for those error
types and then use at least a short `.map_err(Replyify::reply)`.

Luckily, the number of different error types we'll encounter are
likely few, the only one which is currently relevant is `NvmlError`.
Even more interesting is that this also lets us introduce some special
handling for some cases in which we don't want to return a `500` but
some other status.