Impl Replyify for Result<_, Replyify> rather than Result<_, ToString>
We recently implemented `Replyify` for `NvmlError`, which is currently
the only thing we ever shove into the `Err` variant of a `Result` on
which we call `Replyify` functions. And this will likely not change too
dramatically in the future.

This offers the opportunity to base the impl for `Result` on `Replyify`
rather than on `ToString`. This brings a number of benefits, most
notably reduced code duplication.