replyify: move Replyify to a dedicated module
As with most "simple, almost trivial" error handling and reporting
utilities, `Replyify` ends up being a hotspot w.r.t. changes. Not
because we uncover bugs, but because we end up adapting the thing to
more and more situations. And so, we are again on the verge of some more
changes. These will likely include impls of that trait for more types
and maybe more overall functionality. And so we decided that `Replyify`
is worth a dedicated module after all.