replyify: require Result::Ok to be a Reply rather than a Serialize
... for the `Replyify` impl

The motivation of introducing `Replyify` was to make transformation into
a `Reply` as simple and short as reasonably possible. Since all our
current end-points return JSON, we naturally thought of handling that in
the `Replyify` impl for `Result`. However, we intend to introduce some
endpoints that do not return JSON in the response body in the near
future (e.g. a `303 See Other` with an empty body).

Thus, we decided to move the responsibility for encoding `Result::Ok`'s
content to the trait's user.