replyify: make Reply type in Replyify explicit
When we introduced `Replyify` originally, we only had a single impl.
As the design was not as modular as it is now, the `Reply` type for that
impl was complex enough that we did not want to write it out. So we
decided to make use of RPITIT. That choice came to bite us, because
currently, it _really_ makes the type opaque, meaning that we may run
into trouble in situations where it needs to match some other type.

In order to prevent this, we decide to make the `Reply` an explicit type
in the trait.