deps: pull in serve's derive feature
We intend to provide some more sophisticated endpoints for exposing the
GPUs' power consumption over time. The measurement results are strucuted
and contain various elements which cannot be disentagled sensibly.

The easiest way to structure them for us is using the `serde::Serialize`
macro on some transfer data type(s).