deps: pull in tokio's sync feature
We intend to allow starting measuring campaigns and receiving updates.
This requires some data to be stored by the service across requests. We
plan to keep that data in memory rather than on permanent storage, which
means we'll have to mutate some state. `std` does provide mutexes and
read-write-locks, but `tokio` offers async variants which are a better
fit for our use-case (aquiring locks in async functions).