energy: add type representing a base measurment
We intend to provide some more sophisticated endpoints for exposing the
GPUs' power consumption over time. We already introduced a `Measurement`
type for this purpose. However, we can generate a `Measurement` only
relative to some base measurement.

This change introduces a type representing such a base measurement with
fns for creating both a base measurement itself and relative
`Measurement`s.