energy: add fns for creating total and relative DeviceData
Although `DeviceData` is primarily meant as a transfer type, we decided
to also use it for longterm-ish storage for base-data. Thus, we end up
with one use where `energy` denotes the _toal_ energy consumption of a
device since boot/startup and one use where it denotes the consumtion
_relative_ to another measurement.

We plan to create new "total" `DeviceData` and create "relative"
`DeviceData` from it later on. This change introduces fns which let you
do that, as well as some convenience fns for wrapping `NvmlError`s.
Since they are meant for use in `energy` only, we decided it was
acceptable for `total_energy_consumption` take the device id as an
argument despite also receiving the `Device` associated with it in order
to construct more expressive context information.