energy: add type for a collection of addressable BaseMeasurements
We intend to provide some more sophisticated endpoints for exposing the
GPUs' power consumption over time. For this purpose, we recently
introduced `BaseMeasurement`, which provides all the functionality we
need for performing a measurement campaign. However, we naturally do
want to allow for multiple campaigns in parallel. For this purpose, we
need a collection of `BaseMeasurement`s.

We could just use some simple map, but we still need some functionality
for creating new indices. Thus, we decided to abstract that behind a new
type which encapsulates the map and provides functionality for creating
new `BaseMeasurement`s with associated indices.

For now, we go with simple integer incrementation. We deem this
sufficient, since obfuscation is not required. If the REST API was not
public, we would also implement some access control instead of relying
(only) on id randomization.

We also did not implement any sophisticated collision handling other
than erroring out. We expect old entries to be removed by the time
`next_id` is wrapping around, otherwise we'd likely run out of memory
anyway.