Add a store for measuring campaigns
We recently introduced `energy::BaseMeasurements`, which allows
generation of new, addressable `BaseMeasurment`s. However, we still need
to make a shared instance accessible to endpoints.

We decided to use a `tokio::sync::RwLock` wrapped in an `Arc` since it
will need to be static. There are probably better solutions for high
contention scenarios, but at least with an async RwLock we're less
likely to stall or burn cycles.

And currently we're using a single-threaded runtime anyway.