Add custom Duration usable as a query string
We intend to provide some more sophisticated endpoints for exposing the
GPUs' power consumption over time. One of them will be a bit special and
take a measurement over a duration specified via an (optional) query
parameter.

In order to receive that parameter, we apparently need something which
`serde::Deserialize`s from a query string (or, rather, its components).