Add Filter for aquiring a lock on the campaign store
We plan to implement an endpoint for health checks. We already
introduced a fn for compiling the data we want, and that includes data
from the campaigns store.

Previsouly, we didn't have any need for getting a read-lock on the
entire store --apart from this being required for aquiring a lock on an
individual campaign. Thus we lacked a `Filter`.