Add CLI options and default values for garbage collection
We intend to implement garbage collection. We'll assume that any
`BaseMeasurement` older than a certain age will not be used any more and
thus are safe to remove.

At the same time, it's not that expesive to keep some `BaseMeasurement`s
in memory, so kicking off the process is likely not worth it unless we
exceed a certain number. For now, we chose a value that corresponds to
the order of 1MB worth of `BaseMeasurement`s. And because
`Option::expect` is not `const` yet, we need to use unsafe...

Future changes will implement this collection policy.