Add a simple garbage collector
We assume that any `BaseMeasurement` older than a certain age will not
be used any more and thus are safe to remove. We also deem the cost of
keeping `BaseMeasurement` in memory rather low. So it's probably not
worth seeking out and removing old items until we reached a certain
number of them.

We trigger the process either on intervals, which duration is calculated
from the age at which we collect items, and whenever we receive a
notificaiton.