Make base URI configurable
The service may be hosted behind a reverse proxy which places the
endpoints in some sub-directory. While the endpoints themselves don't
care, we do generate a redirect in one of them. Naturally, that redirect
still needs to point to the correct URI.

This change introduces one possible solution to the problem by making
the base URI configuration via a CLI option. If placed behind a proxy,
this option can be used to adjust generated redirects.