Tweak the release profile
The application we develop(ed) is a small, rather simple service which
sleeps most of the time and does only do a fairly small amount of work
every now and then (e.g. when servicing a request).

Fancy optimizations are always nice, but in our case we don't expect too
much load. If anything, we're probably more interested in response time
rather than throughput. And that might be more affected by caching than
by actually running code. Thus, we decided to omptimize for performance
and size for the time being.

Though some of the libraries we use may be complex, the application
itself is not. Even if we ever run into a panic, we can likely take an
educated guess w.r.t. the responsible component. Or we can just carry
on, since the service will probably never be mission critical. We thus
decided that stripping the binary and not generating backtraces is
acceptable, which lets us reduce the binary's size considerably.

And because the program is not that big anyway, we can also enable LTO
without inflating compile times unreasonably.