Add fn for initializing a logger
We currently initialize the logger directy in the main fn, and it's
more of a mess than it needs to be, for various reasons.

We decided that it's best to move this mess out of main, as it's
something that can easily be isolated from all the other setup. And
while we're at it, we took the opportunity to properly implement
increasing verbosity from a base level, which may be specified via the
environment. Sadly this ended up being another mess, since `log` doesn't
provide an easy way to increment a `FilterLevel`.