config: move declaration of default values into dedicated module
We intend to allow configuration via a config file. Using `serde`, we
can implement this by simply declaring some configuration struct and
have it derive `serde::Deserialize`. We also decided to use these same
structs with `clap`'s "derive" feature.

We prefer not to clutter `main.rs` with those config structs. Instead,
we'll put them in a separate `config` module. This change prepares this
module, taking the opportunity to move the configuration defaults.