Add fn for constructing a stream of TCP connections
We plan to support accepting connections via multiple addresses.
Unfortunately, `warp` doesn't come with support for this out of the box,
unless you start constructing multiple `warp::Server`s.

However, `warp::Server::run_incoming` does allow accepting connections
from a `futures_util::TryStream`. This change adds a fn for constructing
such a stream. Due to lifetime- and pinning- requirements as well as a
missing async equivalent to `std::iter::from_fn`, it turned out a tad
more convoluted than we'd like.

Since we'll likely end up using this `fn` in a few branches depending on
the source of addresses, we went for a trait object. A generic fn would
have meant (binary) code duplication, and we decided that we'd rather
life with the expense of dynamic dispatch during startup.