config: add type representing an address to listen on for connections
We intend to allow configuration via a config file. In the case of
network configuration, this will allow for more fine-grained control: on
the CLI, we can only provide one port and a set of IPs to listen on, if
we don't want to introduce some cumbersome, special CLI scheme. However,
in a TOML file we can just have an array of maps.

This change introduces a type allowing to express the latter while
supporting the CLI use-case. The type's deserialize method will allow
parsing configuration either with an explicit port for a given IP or
using a default port, potentially from some other source. We also supply
a `FromStr` implementation which will make it much easier to use with
`clap`'s derive macros.