config: add type representing GC configuration
We intend to allow configuration via a config file. Using `serde`, we
can implement this by simply declaring some configuration struct and
have it derive `serde::Deserialize`. We also decided to use these same
structs with `clap`'s "derive" feature.

This change introduces a struct for configuration of garbage collection,
which impls both `clap::Args` (for the CLI) and `serde::Deserialize`
(for config file parsing). An instance will, at some point, be used as
a member of a `Config` struct which will hold all configuration which
may come from a config file.