config: add general Config container
We intend to allow configuration via a config file. Using `serde`, we
can implement this by simply declaring some configuration struct and
have it derive `serde::Deserialize`. We also decided to use these same
structs with `clap`'s "derive" feature.

This change introduces a struct for all configuration which may be
defined in a config file. It impls both `clap::Args` (for the CLI) and
`serde::Deserialize` (for config file parsing) and integrates all the
topic-specific configuration structs we recently introduced.

In a TOML, each of those will be constructed from a section. For
CLI-parsing, all options are used as-is. Since `clap` does unfurtunately
not support flattening `Args` in `Args`, this required some hand-crafted
impls for `clap::Args` and `clap::FromArgMatches`. We could have avoided
this if we derived `clap::Parser` instead, but that sould obstruct
adding additional options such as `--config`.