Retrieve general configuration via config::Config
We intend to allow configuration via a config file. For this purpose, we
recently introduced a struct `Config`. It enacapsulates all config we
may choose to read from a config file and impls `serde::Deserialize`. On
top of that, we also made it impl `clap::Arg` with the specific goal to
use it also for retrieval of command line options.

This change replaces the existing builder (or, most of it) with an
invocation of `Args::augment_args_for_update` and reads that config.
Currently, we do not yet use that config for anything. Future changes
will migrate away from hand-picking options in various places to
extracting values from the new `Config` struct.

Since we already plan to optionally source configuration from a file
and _then_ from the command line, we use start out from an explicitely
initialized default rather than `FromArgMatches::from_arg_matches`.