util: add fns for deserializing Durations provided numerically
We intend to allow configuration via a config file. Using `serde`, we
can implement this by simply declaring some configuration struct and
have it derive `serde::Deserialize`.

Unfortunately, for `Duration`s it's not that simple. In the config file,
we just want to specify a number of seconds or, in other cases, a number
of milliseconds. To no-ones surprise, this is not how `Deserialize` is
implemented for `Duration`. Luckily, serde does allow supplying an
alternative deserialization method in the form of a fn.

This change introduces such fns.