Make incoming_from consume an IntoIterator
When we originally introduced `incoming_from`, it had to accomodate two
different `Iterator` types. Because we wanted to safe a bit on binary
size, we decided to use dynamic dispatch.

The need to accomodate for differently types sources of `SocketAddr`s
vanished permanently. Thus, we can do away with the dynamic dispatch.