util: add fn for deserializing an Option<Duration> from milliseconds
For configuration parsing, we recently introduced some utility fns to be
used with `#[serde(deserialize_with = ...)]`. We did so because we
wanted the data in the actual type we need, rather than in a proxy type
which happens to deserialize, but needs some conversion afterwards.

The same can be said about types we process as part of requests and
responses. One such instance would be the `DurationParam` which
encapsulates an `Option<Duration>`.

This change introduces a `fn` to use with `deserialize_with = ...`. A
future change will wire it up.