util: add fn for serializing an Duration as number of milliseconds
For configuration parsing, we recently introduced some utility fns to be
used with `#[serde(deserialize_with = ...)]`. We did so because we
wanted the data in the actual type we need, rather than in a proxy type
which happens to deserialize, but needs some conversion afterwards.

The same can be said about types we process as part of requests and
responses. One such instance would be the `energy::Measurement` which
contains an `u128` which really should have been a `Duration`.

This change introduces a `fn` to use with `serialize_with = ...`. A
future change will wire it up.