Make nvml in fn main a &'static Nvml rather than an Arc<Nvml>
`warp::Filter`s need to implement a number of traits, leading to any
`Fn`s used in them to be at least `Clone` and `'static`. For some of
them, we need to get hold of _the_ `nvml::Nvml` instance. Because we
felt this was the most simple option back then, we decided to just wrap
it in an `Arc`.

However, what we really always wanted was a `&nvml::Nvml`. A reference
has the benefit that it impls `Copy`, which will safe us from using
`Clone::clone` in an awful lot of places. A `&'static nvml::Nvml` also
allows for `'static` `nvml::Device`s, which was not viable since the
device's lifetime is restricted by the `nvml::Nvml`'s.

This change makes `nvml` in `fn main` a `'static` reference, which we
achieve through a `static` `OnceLock`. As `OneLock::get_or_try_init` is
not yet stable, we place the initialization outside the closure.

This change also applies some necessary changes due to the changed type,
but it does not eliminate now unnecessary explicit cloning. That will be
done by future changes.