Add Filter for deriving a nvml::Device from a path param
Previously, due to lifetime issues we needed to retrieve devices in an
endpoint using a reference derived from an `Arc<nvml::Nvml>`. However,
we recently made `nvml` in `fn main` a `'static` reference. This now
allows to construct a `warp::Filter` which transforms a path paremeter
into a `nvml::Device<'static>` that we can then forward to the actual
end-point function.

Future changes will migrate all relevant end-points to this new filter.