energy: add type representing a base measurement for a single Device
In the `energy` module, we ended up storing `nvml::Device`s's ids since
that was the only viable option. And because we were lazy, we ended up
re-using `DeviceData`, which was initially meant solely as a transfer
type.

However, due to recent changes we can now use `'static` `nvml::Device`s.
This is a great opportunity to both eliminate an awkward lookup of
`Device`s and properly separating base measurements for a device from
the relative measurement we actually send out to clients.

This change introduces a new type representing the base measurement,
complete with a fn to compute the relative measurement expressed through
the already existing type. Future changes will migrate over to that one
and eliminate any motion of a "total" `DeviceData`.

We deliberately did not make use of existing conveinece fns we
introduced for `DeviceData` since especially relevant device id
information is attached differently in the new code.