energy: migrate BaseMeasurement from DeviceData to BaseDeviceData
In the `energy` module, we ended up storing `nvml::Device`s's ids since
that was the only viable option. And because we were lazy, we ended up
re-using `DeviceData`, which was initially meant solely as a transfer
type. However, due to recent changes we can now use `'static`
`nvml::Device`s, and we already introduced a new `BaseDeviceData` making
use of that.

This change migrates `BaseMeasurement` over to that new type, making
`nvml` an unused parameter in `measurement`. Some future change will
eliminate that parameter.