energy: make DeviceData a pure transfer type (again)
Although `DeviceData` was meant as a transfer type only, we ended up
using it also for storing base measurements. This lead to two
semantically different uses: "total" and "relative" device data, the
latter being what we send to clients.

However, we recently introduced a new type to fill the role for base
measurements. Thus, we no longer need to make this distinction, nor do
we need or want the associated fn for creating "total" `DeviceData`.
Consequently, we also don't need or want a fn for "transforming" any
"total" `DeviceData` into "relative" ones.