util: add a Reject type for failure to retrieve an nvml::Device
Previously, due to lifetime issues we needed to retrieve devices in an
endpoint using a reference derived from an `Arc<nvml::Nvml>`. However,
we recently made `nvml` in `fn main` a `'static` reference. This now
allows to construct a `warp::Filter` which transforms a path paremeter
into a `nvml::Device<'static>` that we can then forward to the actual
endpoint function. However, for this we also want a `Reject` that we can
feed into `warp::reject::custom` for a `500 Internal Server Error`.