Make campaigns in fn main a static reference rather than an Arc
`warp::Filter`s need to implement a number of traits, leading to any
`Fn`s used in them to be at least `Clone` and `'static`. For some of
them, we need to get hold of the campaigns container. Because we need to
mutate the container in some end-points and during garbage collection,
we put it in an read-write-lock. And because we felt this was the most
simple option back then, we decided to just wrap it in an `Arc`.

The `campaigns` container is infrastructure that needs to life
throughout the entire run-time. Ideally we would use a `static`
variable, which would also let us use "regular" lock guards rather than
the thankfully availible but still more expensive owned variant.

Unfortunately, `HashMap` doesn't have a `const` constructor, which
obstructs one for `BaseMeasurements`. Wrapping it in a `OnceLock` is one
possible solution. And just as we did with `nvml::Nvml`, we can use that
to create a `'static` reference which we can use to make the read-write
lock available in the places we need it without the overhead of an
`Arc::clone`.