Use single Filter chain with property name for device end-points
We have a number of end-points which represent a property of a device.
Previously, every single one of them is implemented via its very own
`warp::Filter` chain. Because we combined them with `Filter::or`, when
a request for anything in `/v1/device/` came in, we have to consider
each of them after another until we found the correct one. This included
looking up the device by its id. This is clearly not ideal.

This change defines a single `Filter` chain with one param for the
device id and a second one for the property. Thus we only need to look
up the device once.