param: add parameter representing names of device properties
We have a number of end-points which represent a property of a device.
Currently, every single one of them is implemented via its very own
`warp::Filter` chain. Because we combine them with `Filter::or`, when
a request for anything in `/v1/device/` comes in, we have to consider
each of them after another until we find the correct one. This includes
looking up the device by its id.

This is clearly not ideal. Rather, we would like to look up the device
only once (and fail only once if the device does not exist).

Taking a step back, every end point in `/v1/device/:id/` represents and
queries some property of a device. If we enumerate those properties, we
can define a single `Filter` chain with one param for the device id and
a second one for the property. This would mean we only need to look up
the device once, just as we wanted. Using an enumeration (instead of,
say, a `String`) also still lets us use `Filter::map` for the final
stage as a `match` in there doesn't need to handle invalid property
names.

This change introduces such an enumeration. A future change will
reorganize the device end-points.