config: add config option for enabling oneshot end-points
Our service supports one oneshot end-point, which blocks the connection
for a duration which, on top of that, is defined by the client. While
this usually isn't an issue with clients connecting directly to the
service, it becomes an issue if a reverse-proxy is involved. Those
usually use a limited number of connections to origin servers. Thus, any
blocking request also may block requests by other users.

We could highlight this in our documentation, possibly leading to some
operators blocking or restricting access to this particular end-point
through the proxy. However, we decided to disable this end-point and
provide a config option for enabling it (and any other future blocking
operations) if desired.

This change only adds the configuration option, both via the CLI and
config file. Future changes will wire this up such that it controls the
availability of the end-point.