Enable oneshot energy measurement end-point based on config
Our service supports one oneshot end-point, which blocks the connection
for a duration which, on top of that, is defined by the client. While
this usually isn't an issue with clients connecting directly to the
service, it becomes an issue if a reverse-proxy is involved. Those
usually use a limited number of connections to origin servers. Thus, any
blocking request also may block requests by other users.

We decided to make the availability of this end-point controllable via
run-time configuration. We recently introduced an config option for this
purpose, which did not yet have any effect. This change introduces a
`warp::Filter` which rejects requests based on that option, and applies
it to the existing oneshot measurement end-point.