Ensure base URI has no query and ends with '/'
We use the base URI to construct URIs for redirects, by simply appending
a relative path --partially because `http::Uri` does not provide any
reasonable API for manipulation of the path component.

If we want to at least somewhat ensure we end up with a well-formed
result, we need to control, or sanitize, the base URI, specifically the
path and query portion. For the latter, the only sensible thing is to
reject any query, as this obstructs path manipulation (and doesn't make
any sense to begin with). For the path, we only really can _add_ things.
We therefore decided that a base_uri should always end with a '/'.