config: add struct for miscellaneous config options
We use TOML as a format for config files, with the structure dictated by
the structure of the `Config` struct. In that struct, most things are
grouped in sub-structs, which translate to sections in the TOML format.

The sole exception to this rule is the base URI, which lives at the top
level. This results in awkward placing requirements, since the only
sensible way to place a string value at the top level is to place it
before any section. We therefore decided to introduce a new section for
micellaneous stuff.

This change introduces a configuration struct for the new section, with
the base URI. Future changes will add it to `Config`.