config: add config option for max-age
We plan to include cache-control directives in our responses. Besides
`no-cache`, the `max-age` directive appears to be applicable in our
case. For that, we need to conjure some sensible value. As a default, we
settled on "15 minutes" since, for example, device ids may change after
a reboot. Still, it's probably a good idea to make this configurable.