replyify: add additional convenience trait
The `Replyify` trait helped reduce boiler plate code considerably. One
of the remaining paint points is transformation to JSON, which requires
a closure since `warp::reply::json` takes a reference.

Back when we introduced `Replyify` we decided not to make it do the
conversion to JSON, since it's sometimes not what we want. Because we
still want to make JSONification more convenient, we now decided to
introduce a new extension trait with a fn which does the conversion to
JSON and subsequent replyification.

We attached the trait bounds for this to the fn rather than the types,
because we plan to add more fns in the future, with possibly very
different bounds.