util: add fn for sanitizing a base URI
From the configuration, we receive an URI which we then use as a base
for constructing other URIs, e.g. for redirects. This URI needs to be
sanitized, and after agonizing over it we decided to perform the
sanitization in the main fn.

Due to future plans, we now decided that we want to sanitize the URI
before it is stored in `config::Config` after all. For this purpose,
this change introduces a sanitization function, which we'll use in
future parsers and deserialization helpers.