util: make deserialization helper sanitize base URI
From the configuration, we receive an URI which we then use as a base
for constructing other URIs, e.g. for redirects. This URI needs to be
sanitized, and after agonizing over it we decided to perform the
sanitization in the main fn.

Due to future plans, we now decided that we want to sanitize the URI
before it is stored in `config::Config` after all. For this purpose,
we recently introduced a sanitization function. This change adapts the
existing `deserialize_uri`, making it sanitize the URI, and renames it
to reflect the new responsibility.