config: use new sanitizing fn for parsing base URI command line option
From the configuration, we receive an URI which we then use as a base
for constructing other URIs, e.g. for redirects. This URI needs to be
sanitized, and after agonizing over it we decided to perform the
sanitization in the main fn.

Due to future plans, we now decided that we want to sanitize the URI
before it is stored in `config::Config` after all. For this purpose, we
recently introduced a new sanitizing parsing fn. This change finally
wires it up as a value parser for the `--base-uri` command line option.