Remove now redundant base URI sanitization from fn main
From the configuration, we receive an URI which we then use as a base
for constructing other URIs, e.g. for redirects. This URI needs to be
sanitized, and after agonizing over it we decided to perform the
sanitization in the main fn in the past.

Due to future plans, we now decided that we want to sanitize the URI
before it is stored in `config::Config` after all. We thus introduced
sanitization in deserilizer and parser helpers, rendering this step
redundant.