replyify: add ResultExt fn for attaching a header to the Ok variant
We plan to include cache-control directives in our responses. As
`cache-control` is a header, we need to attach it after replyification.
However, we generally only want to attach it for successful results, not
for errors.

We don't want to do exessive `Result::map`-ing in closures (again). To
avoid this, we ideally want some fn associated to `Result`. And that's
what we have `ResultExt` for to begin with.