replyify: add fn for attaching a cache-control directive to ResultExt
We plan to include cache-control directives in our responses. As
`cache-control` is a header, we need to attach it after replyification.
However, we generally only want to attach it for successful results, not
for errors.

We recently introduced `ResultExt::with_header` to attach headers to
successful responses. But since we're interested specifically in
`cache-control`, we may as well introduce a convenience fn for this
case.