Include cache-control directives in responses for GET end-points
With our service, we do have some end-points which should not be cached.
We could ask the operatos of any reverse proxies to deactivate caching
for our service, or ask clients include a `no-cache` directive, but
including one ourselves is probably the best idea. Besides, some of the
responses we generate can actually be cached, and differentiating
between them is easier on our side.

For example, device info is generally cachable, but not desired
specifically for the power-usage end-points. Measurements, on the other
side should never be cached for obvious reasons. For ping and health we
decided to not have them cached, since the whole point of those is to
get an idea of the origin's server's state.