Dockerfile: add initial Dockerfile bundling the service
Some users will want to run this service in docker. For their
convenience, we provide a Dockerfile. We base it on `nvidia/cuda`. That
image is arguably bloated considering all we want is `libnvml`, but it
will do for now. Its somehow still smaller than if we used `ubuntu` as a
base for the runtime...