config: do create additional sock for listening on IPv4 by default
On Linux, if the `IPV6_V6ONLY` socket option is off, listening on `::`
will also catch IPv4-connections to the same port. Thus, the second bind
would fail with `EADDRINUSE`. Unfortunately, we were oblivious to this
detail.

For now, we decided to listen on `::` only by default. For systems where
this is not sufficient, users can simply supply additional config either
via a TOML or the command line. In the future, we may add additional
config options for controlling the `IPV6_V6ONLY` option of the listening
sockets.