main: fix "about" string displayed with `-h` and `--help`
We impl `clap::Args` for `config::Config` manually. We augment a
`clap::Command` by applying the augmantation of each sub-config, which
is generated through a derive macro. And that will also update the
"about" and potentially the "long about" from the doc of those structs.
Which is obviously not what we want.

Unfortunately, there's no way to tell the derive macro to just leave
those things alone. Thus, we need to clean up after the fact.