health: introduce struct encapsulating data for performing health checks
We currently perform a health-check via the `health::check` fn, which
takes as arguments all the things it needs to perform such a check. It
is called in an end-point which, of course, needs access to all of those
inputs. And while most of them are `Copy`, we did need an extra binding
for `oneshot_enabled`. As we add more things to the health-check in the
future (such as the driver version the service was started with), this
may get a bit unwieldy.

We therefore decided to encapsulate all of this in one struct `Checker`
(with the exception of the campaigns, since they are behind an `RwLock`)
and make `check` an associated fn of that struct. This change introduces
this struct, but preserves at least the old interface, so that we can
migrate in multiple steps.

We could have made `Checker` impl `Copy`, but we already intend to
tracking the driver version we started with, and that won't be `Copy`.
We also made `Checker::new` return a `Result` because retrieval of some
data we'll need to do during setup in the future could potentially fail.