Use new health::Checker for health-check end-point
Previously, we performed health-checks via the `health::check` fn, which
takes as arguments all the things it needs to perform such a check. It
was called in hte health-check end-point which, thus needed access to
all of those inputs. This would have gotten a bit unwieldy when adding
more things (such as the initial driver version) in the future.

We thus recently introduced a type encapsulating most of this data with
an associated fn for performing a health check. This change now migrates
the health end-point to this new helper.