health: include device names in health info
Sometimes, maintenance on the underlying machine breaks the service:
we'd still get some basic info, but error out on any attempt to get a
`Device` handle. Thus, the health end-point would still indicate that
everything be fine, while attempting to start a campaign would lead to
an error when trying to retrieve `Device`s.

This change now includes device names in the health info. Fetching them
during a health-check entails querying for all `Device`s, which means we
then _hopefully_ error out the same way we would when starting a new
campaign or querying other device info. We choose the name because that
appears to be the only info somewhat guranteed to be present for all
devices (we gloss over non-utf8-device names for now).