client.py: add dedicated data types for measurements
Currently, various fns return measurements as the JSON we receive from
the server as is. In real production code, we would likely want to have
dedicated data types for measurements.

`client.py` is not meant for production but as an example/demonstrator.
Still, data types we use here should reflect the structure we expect the
JSON to have. Thus, we decided to introduce data types at least for the
primary use case.

This change introduces these data types as well as a fn performing the
conversion from JSON. Future changes will make `MongeuClient.oneshot`
and `Campaign.get` return `Measurement`s.