client.py: add fn for printing Measurements
We currently print measurements as we get them from the `MongeuClient`
or `Campaign`. In the future, we will receive `Measurement`s instead,
and those are not printable as is. We could define `__format__` for
them, but there's no obviously correct way to format a measurement.

With campaigns, a roughly tabular form seems appropriate. Thus, we
decided to introduce a fn for printing measurements as one line of such
a table.