client.py: rename class Client to MongeuClient
When naming entities, we try to not repeat terms occurring in the name
of the module the entity lives in. Thus, in a module named `mongeu` we
would not name a type `MongeuClient` but simply `Client`. A user would
then either import `mongeu` and spell the client type `mongeu.Client` or
rename the type as appropriate, e.g.

    from mongeu import Client as MongeuClient

However, while `client.py` can be adapted to serve as a python module,
it currently isn't. In the current situations, naming the client class
`MongeuClient` turned out to be less confusing to users.